Add files via upload
update  models

+++ models/menuItem.js
@@ -15,7 +15,7 @@ module.exports = (sequelize, DataTypes) => {
   }
   MenuItem.init({
     name: DataTypes.STRING,
-    price: DataTypes.FLOAT,
+    price: DataTypes.INTEGER,
     description: DataTypes.TEXT,
     imageUrl: DataTypes.STRING,
     storeId: DataTypes.INTEGER,
